Add script helper to sync the devstack hosts to /etc/hosts
This script iterates through the devstack hosts on serverstack and adds
the public IP addresses to the /etc/hosts file so that the user can use
the host names rather than IP addresses.  Note, it needs concommit
changes in the ~/.ssh/config files to manage the changing IP addresses
and usernames.